Add support for replace_imports_with_any for global wildcard
Summary: I realized Fixit has [`ignore_missing_imports` in the root level of the mypy config](https://github.com/Instagram/Fixit/blob/main/pyproject.toml#L106), and saw that this wasn't being properly migrated when we run `pyrefly init`. I dug into this and it looks like the reason is because we are only migrating `ignore_missing_imports` when it's in a subsection (e.g. `mypy-dir1`). This diff adds support for when it's in the root level, and converts it to a wildcard

Reviewed By: stroxler

Differential Revision: D78490005

fbshipit-source-id: ba124871e4ad3267a5144b1394bdeea9319ea910