Make InitVar/NamedTuple consistent
Summary:
Previously:

```
class C(NamedTuple): fld: InitVar[int]

for x in C(1):
    reveal_type(x, Never)
```

That is wrong, so instead make it Any. There are three choices here, and Pyrefly, Pyright and Mypy all pick differently. If we do want to go with Any, we should raise an error earlier. But I leave this for people who understand this better than I do.

Reviewed By: yangdanny97

Differential Revision: D78543661

fbshipit-source-id: 53e5b736537464057152a6d6167183ba76cc825a