Warn on deprecated overloaded calls
Summary:
I noticed while reviewing D78525047 that we also don't handle overloaded functions. With this change, we warn on overloaded functions that are fully deprecated (i.e., the implementation is marked deprecated). We don't yet handle cases in which a single overload is deprecated.

I had to do a slightly hacky thing to propagate the is_deprecated flag, because trying to use the function metadata from the overload implementation rather than the first signature broke a bunch of conformance tests for some reason.

Reviewed By: ndmitchell

Differential Revision: D78527825

fbshipit-source-id: 977b8c0c220b3e1871fc854fa7c6c52632968046