Add missing test
Summary:
I think there was a phase where we were shipping features on
synthesized classes using only conformance tests to validate.

This is really bad for devx; I hit some errors on a refactor, and
it takes many, many minutes to get the feedback versus reasonably
quick turnaround from unit tests.

This diff imports the case where my refactor had an issue as a unit
tests to prevent a repeat of the experience.

Reviewed By: grievejia

Differential Revision: D78464144

fbshipit-source-id: b98f2c4139977eb6d2d60008cc462809006f5212