Let's not define `to` to be `self.type_order`
Summary:
I rely quite a bit on code search for refactors, and having a variable
named `to` that is `self.type_order` messes with this.

I don't think the smallish reduction in verbosity is worth that cost;
in addition, while `to` is a reasonable acronym for "type order" at
first glance, it's also really weird to see it in subset code because
it kind of reads like the word "to" which might indicate some kind of
assignment target (the "to type").

Reviewed By: ndmitchell

Differential Revision: D78566083

fbshipit-source-id: 57f4279931414501ade7238cc27bf977d0edff15