Clean up RawClassFieldInitialization
Summary:
At this point, the original datatype is an implementation detail of
the messy logic in calculate_class_field, which I'd like to refactor
but that's not yet in scope.

Let's use a new datatype to represent the class body scope information
that doesn't have extra nonsense in the representation, and make the
old type a private helper datatype in `class_field.rs`.

At this stage, the bindings-side information roughly makes sense and
behaves as I think it should.

Reviewed By: grievejia

Differential Revision: D78464149

fbshipit-source-id: 18464889c03e2cb03630c2f1e64c36c51be176c1