Eliminate strange special-form case (probably a bug?)
Summary:
I was extremely confused by the block eliminated in this diff,
it makes no sense to me: why would I want to, for example, treat
the type of an attribute named `tuple` in a synthesized class
as automatically having the builtin tuple type? That's just wrong.

It was introduced, with no discussion in the diff comments or
review notes, by D70004838 which implemented the functional form for
`TypedDict`. I cannot imagine any reason why this would be intended or correct,
and no tests fail if I remove it, so I'm just removing it.

My best guess is that it was an artifact of debugging based on a faulty
understanding of something or other that we accidentally slipped into
trunk, which happens.

Reviewed By: migeed-z

Differential Revision: D78464154

fbshipit-source-id: 26851c326b2d71f765350428e40913a9cdd0e343