Fix `key-already-exists panic` when checking nested comprehensions that are incomplete (#693)
Summary:
Resolves https://github.com/facebook/pyrefly/issues/455
An incomplete nested list comprehension `[a for [a for` is parsed into two `ListComp` expressions with the same `iter`, which is empty. I assume this is correct because there are no objects to iter over. The panic occurred due to inserting `comp.iter` binding twice. If two iters have the same range, then it might be safe to consider them as the same expression. Therefore, it is safe to overwrite the binding.

But I am not sure if this is the best option. Alternatives might be:
- check existing bindings before inserting and ignore them
- or produce some error for the user.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/693

Test Plan:
Added two tests to pyrefly/lib/test/simple.rs
1. test_nested_list_comp - checks nested list comp type inference
2. test_incomplete_nested_list_comp - checks no panic for incomplete nested list comp

Reviewed By: ndmitchell

Differential Revision: D78576459

Pulled By: migeed-z

fbshipit-source-id: d5cf0e03a64258389113cc8d30bb335504434b7b