Eliminate extra `AnswersSolver` impl
Summary:
Currently `attr.rs` has three(!) implementations of `AnswersSolver`, which
is pretty confusing.

Two of them kind of make sense: we have the main one, and then another related
specifically to lsp logic that is trying to work around aspects of the main
`attr` api that don't map trivially to lsp logic, and there are some lsp specific
data structures defined in between (why do these things live together? Because the
lsp code needs access to private information that we really don't want to expose
to all of Pyrefly).

But the third one did not make much sense to me. I'm just moving the `__bool__`
helper up to the main implementation block.

Reviewed By: yangdanny97

Differential Revision: D78572826

fbshipit-source-id: c7b259b2c1cde1c784d9e6cd401955f6d4ba0cb7