Processing Functions parameters
Summary:
Processing functions parameters filling additional fields for function definitions and creating the corresponding variable declaration. Note that we need to make sure the top_level_declaration facts for these variable is the function if needed, so, I added an optional field to the DeclarationInfo struct, to be able to override the default top_level_declaration if needed.

Adding the type information is a follow up

Reviewed By: stroxler

Differential Revision: D78555436

fbshipit-source-id: 51dda8aed19e033c6fdc06062eb64879db0949a3