store docstring as range
Summary:
Docstrings are currently stored in the bindings table raw. Although they're in an `Arc`, they still require an allocation. By only storing the range, we can avoid this extra storage.

This diff seems very large, but a lot of it is a rename. I could've planned better and done this rename of `docstring` -> `docstring_range` in a followup diff, but that is difficult now. Hopefully splitting D78555843 and D78561914 make it appear small enough to review.

Reviewed By: ndmitchell

Differential Revision: D78557762

fbshipit-source-id: 5e48f49a84d5ba0f02ec13b148f8730636978165