Raise error on variable use before write (#666)
Summary:
Closes https://github.com/facebook/pyrefly/issues/15, https://github.com/facebook/pyrefly/issues/248

This change adds additional checks when building bindings, so that
variables that are used before initialised are correctly flagged as
errors.

Currently, we error on uninitialised variables by checking their flow
style. If the flow style is uninitialised or partially uninitialised, an
error is generated. When no flow style information is obtainable (which
is the case for use before write), a default "other" flow style value is
returned.

In this diff, we perform additional checks before returning a default
"other" value. By looking up the scopes for static information for
variables, we can correctly detect if a variable is declared in an upper
scope or in the current scope, and return "uninitialised" flow style
when the variable is used before write.

When doing so, we need to consider the special case for type usages,
so that name resolutions in types do not fail.

Help needed:
There's a failing test that I need some help: `test::flow::test_assert_not_in_flow`

```
testcase!(
    test_assert_not_in_flow,
    r#"
from typing import assert_type, Literal
if 0.1:
    vari = "test"
    raise SystemExit
assert_type(vari, Literal["test"])   <-- Shall we raise an error here?
"#,
);
```

Pull Request resolved: https://github.com/facebook/pyrefly/pull/666

Reviewed By: yangdanny97

Differential Revision: D78496802

Pulled By: stroxler

fbshipit-source-id: 5c13e661f90b6b2c701e500f4aec9745b8a4a1c2