Move `ensure_mutable_name`
Summary:
It arguably makes sense for this to live in `stmt.rs` because
it's only used on mutation statements.

But I think the more significant factor is that it's tightly
coupled with how name lookups work, so for architectural reasons
I think we're better off consolidating the lookups all in one
place.

I ran into this because I did a refactor of the name lookup logic
and got tripped up on this function; it would have been easier to do
the refactor if we'd had all name lookups grouped together.

Reviewed By: yangdanny97

Differential Revision: D78677966

fbshipit-source-id: b36129ad8c3d929716a4add13377e392116a734d