Push context into typed dict constructors
Summary:
Danny added this context propagation for class constructors in D73866382. We can do the same thing for typed dict constructors.

I'm really only adding this test (and fixing it) to make sure I don't break it in later diffs. I don't think this is a very important contextual typing use case. The more important case is a dict literal expression with a TypedDict hint, which already works.

Reviewed By: ndmitchell

Differential Revision: D78677083

fbshipit-source-id: 3f20e573cd21765017e224d4c777849d75902ff6