Push context through assert_type and reveal_type
Summary:
Both of these functions have the form `[T](x: T) -> T`. In general we should be pushing an expected return type through to the argument for functions like this, but that is still TODO.

Adding these special cases because (a) general support won't help for these special calls anyway, (b) it's useful to be able to add this code without changing behavior, and (c) we can prevent "contained" vars from leaking out in the type checker output (printing `@_`) in many cases.

(Aside, for (c), we might consider deferring the printing of errors for reveal_type and assert_type until contained vars are solved.)

Reviewed By: ndmitchell

Differential Revision: D78675560

fbshipit-source-id: 78482e15c481c44a67bda2f9675c69877f0c0be1