Let `resolve_named_import()` declare&use its own gas
Summary:
In all but one caller of `resolve_named_import()`, we invoke the function with the `INITAL_GAS` constant. The one remaining caller tries to re-use the same gas with `key_to_intermediate_definition()` which I think is not desirable (we never restore gas in both methods, so if one function uses up the gas the other lookup will fail even if the gas cost is reasonable.

Let's rearrange the code so that `resolve_named_import()` uses a dedicated gas var.

Reviewed By: SamChou19815

Differential Revision: D78689925

fbshipit-source-id: b5aef7b0661fa062bcfd9e15fbe00d4246233e6c