Defer self_arg creation until callable_infer
Summary:
Passing down an owned type instead of the &Type in CallArg is necessary so we can substitute Quantified with fresh Vars here, as part of my ongoing work to separate constructor instantiations in overloaded calls.

Also, this clarifies the code IMO. In particular, we can see that self_arg is never an Expr, so it becomes clear that forcing exprs to types in self_arg is a noop.

Reviewed By: rchen152

Differential Revision: D78691966

fbshipit-source-id: 3ddd45b6c52585fb2593526bf0731e241967d775