Remove gemfile (#718)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/718

Just seeing what happens if I remove this.

It was copied over from flow's website and has not been changed since, and I am confused
about why it's here. The docs make no mention of any ruby tooling.

If CI passes with this removed, I'm almost certain it is not needed (the website built fine without it)

Reviewed By: SamChou19815

Differential Revision: D78703559

fbshipit-source-id: c0237e30415ee38abcea2c0cedbc1dfae2eb2436