Try turning on rust-cache for the `pyrefly` workflow (#720)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/720

I'm not totally satisfied that I understand perfectly how this works, it's
at least possible that our override of `CARGO_HOME` will impact it on
windows.

But it's worth trying. The exact config snippet I pulled from a buck
example.

I think the pyrefly workflow accounts for the majority of our rust build
time, so I think that if this works, we should see some actual results
in both turnaround time and compute use.

Reviewed By: ndmitchell

Differential Revision: D78710986

fbshipit-source-id: 515938e74d7de6dcc6687065f9dc90ed0d3755fb