Consolidate `lookup_name` to one function
Summary:
We had a `lookup_name_impl` with two wrappers that each elimninated
one argument. But the uses of these helpers weren't really consistent
(e.g. some of the non-pinning cases were getting the wrong kind of Usage
which is confusing).

It seems better to just have one and make the callers actually specify
what they are doing directly; there are few enough callsites that I think
we actually get less boilerplate, not more, by eliminating the wrapper
functions.

The only place a helper funciton really seems merited is in `target.rs`
where we're using lookups to just check whether a name even exists (as
part of narrowing), which also behaves a little weirdly so the helper
function is handy as a way to explain what's going on.

Reviewed By: ndmitchell

Differential Revision: D78682203

fbshipit-source-id: 9c31ed8b7e6a186a067cf834d7261c2680e9854f