Consolidate lookup logic under `ensure_name`
Summary:
This diff (and the stack of preceding refactors) is a response to
a diff comment noting that the arguments for `ensure_name` were becoming
less obvious.

I realized part of the issue was that we were probably doing too much
on the caller side, I think it makes much more sense for `ensure_name`
to handle the lookup-name calls.

This diff makes that change.

Reviewed By: ndmitchell

Differential Revision: D78682562

fbshipit-source-id: 3ef4f1567a445073bc31839082a104ecbd7e10b8