Rework the summary line
Summary:
Some user feedback was that the summary line was quite complex. It used to always say:

```
 INFO errors shown: 4,758, errors ignored: 8,123, modules: 1,946, transitive dependencies: 3,861, lines: 2,672,185, time: 1.42s, peak memory: physical 1.7 GiB, allocated 926.7 MiB, active 1 GiB
```

That's a lot of information. Useful information. And you could hide it all with `--no-summary`. This diff gets rid of `--no-summary` in favour of defaulting to:

```
INFO 27 errors (8 ignored)
```

And allowing `--summary=none` to hide the summary, or `--summary` to display:

```
 INFO 27 errors (8 ignored)
 INFO 1 module (9,225 dependent modules, 1,921,458 lines); took 1.69s (checking 1.03s, listing 0.60s); memory (physical 19.1 GiB, allocated 3.3 GiB, active 12.1 GiB)
```

Downside is we break compatibility with `--no-summary`. But that seems OK pre-beta, and given people who were passing no summary might really have wanted what is now the default behaviour, that doesn't seem terrible.

At the same time I decreased the threshold for displaying intermediate times, now they are off by default its more reasonable to show them more readily.

Reviewed By: kinto0

Differential Revision: D78733609

fbshipit-source-id: 80b7e4ff0a99e25b7b28d1250a5b51e4f5050fc6