Fix a small bug in assignability checking
Summary: Fixes a bug where we didn't take readonly-ness into account. Also fixes a comment that described the assignability relationship backwards. Confusingly, the spec and the conformance tests flip the names they use for the "got" and "want" TypedDicts when describing the relationship (i.e., the spec says B is assignable to A if <blah>, whereas the conformance tests say A is assignable to B if <blah>), which may be why we got this wrong.

Reviewed By: samwgoldman

Differential Revision: D78776806

fbshipit-source-id: 12345d101adbbc57979acb4aa455a35973697471