Don't expose Attribute to named tuple code
Summary:
I'm at least toying with the idea of trying to eliminate or dramatically
refactor internals of the Attribute datatype, which I don't think really
represents the cleanest interface.

My rationale is that the interface of attribute is one of *actions*,
not *things*, and the fact that we treat attributes as a "thing" forces
us to represent the cross product of all possible actions in a data
structure; I'm not sure this gains us much.

Just to explain my thinking a bit more: a good example of how Attribute can
complicate our life is the way that read descriptors vs data descriptors
actually "descriptor-ify" disjoint actions - a data descriptor only changes set
and delete actions, whereas a read descriptor only changes get, but having a
monolithic Attribute datatype forces us to represent all this in a big enum,
with a lot of the logic also duplicated with ClassField.

I'm not yet sure whether or if I will do this, but even if I don't I'm
pretty confident it's a good thing to expose Attribute in as few places as
possible to give us flexibility, because we already know we have bugs related
to unions that will require major changes to how we work with Attribute at
some point.

This diff trivially hides `Attribute` from named tuple code by just doing the
lookup in one shot, without exposing the intermediate `Attribute` datatype
which is really just an implementation detail here and doesn't need to be
in the public API.

Reviewed By: rchen152

Differential Revision: D78767920

fbshipit-source-id: 563e41776f202690bf02aaa015fb923140bf342b