Eliminate Attribute exposure in `subset.rs` / `type_order.rs`
Summary:
This diff continues hiding the `Attribute` datatype, which I want to have as at *most*
an implementation detail in the `class_field` / `attr` module interaction - not exposed
to the rest of Pyrefly - by making a method in `attr.rs` that handles all of the logic
for checking protocol subset at one attribute name without exposing the intermediate
data.

This gives us the option of potentially moving away from `Attribute` as the contract
between the `class_field` and `attr` modules, because it is now an implementation
detail only used in those two places.

I think this may be useful because I'm not convinced representing `Attribute` as
a closed-form datatype (as opposed to just directly dealing with *actions* on attributes)
is a net positive.

Even if we wind up keeping `Attribute`, I think hiding it from most of the code
is cleaner and more modular

Reviewed By: ndmitchell

Differential Revision: D78767919

fbshipit-source-id: 3efcef452dcd2596ffbcf34542d2d791903b4046