Rename some methods
Summary:
I don't love it when we wrap a method without changing anything about it and
rename it; let's eliminate that from `type_order`.

While we're at it, change `attr` to `attribute` to emphasize that it operates
on `Attribute`; in a large fraction of our methods `attr` is more like a name,
and in fact I want to make a new `is_attr_subset` that is name-level in order
to hide `Attribute` from the subset logic. The rename will make it easier to
do so.

Reviewed By: rchen152, ndmitchell

Differential Revision: D78767922

fbshipit-source-id: 1276104999d0ad58500e6204ae9b6de973045770