Simplify Narrowing
Summary:
At the time I initially implemented attribute narrowing, we were considering whether
to implement sound narrowing where we complain for narrowing on descriptors.

Having had more time to think about the big picture, this seems like a not very useful
idea to me. In order to be useful, the type checker already has all kinds of unsoundness:
  - we assume you can safely narrow globals and nonlocals, this is never true but needed
    pragmatically because the kinds of issues that can pop up *almost* never do, and narrowing
    is too prolific to be 100% safe
  - we allow index narrowing; you can always implement a `__getitem__` that returns random
    results, so in theory this makes no sense if users write highly non-idiomatic code, but
    again that is very rare, and the ability to narrow an index is extremely useful

The descriptor narrows are almost exactly analagous to `__getitem__` narrows: idiomatic
descriptor code should be idempotent, and in hindsight I think we should just rely on this.

Eliminating the tracking of descriptors simplifies the narrowing lookups quite a bit; this
diff removes the complexity, in the next diff I'll actually delete the Narrowable type (I
found it easier to do it in two stages, logic first and then types)

Reviewed By: ndmitchell, yangdanny97

Differential Revision: D78767921

fbshipit-source-id: 23128fe18352e76c8edba310c738b086c2d65d80