Rename as_bound_function() method on BoundMethod to as_function()
Summary: Depending on the day and my mood, I think of the underlying function in a bound method as either "bound" (because of the action of it being "bound" to the self object) or "unbound" (because it's the "unbound" form of the method), and half the time I get confused about what to_bound_function() does. IMO to_function() is sufficiently clear and side-steps any confusion about bound/unbound terminology.

Reviewed By: ndmitchell

Differential Revision: D78789367

fbshipit-source-id: ab2c4e2684d0a0d56df0c9bd27a363f8543aaa64