Simplify metaclass __call__ logic for class constructor calls
Summary:
We don't need to track `overall_ret` here -- the only information we care about from the metaclass call is that the return type is compatible with the class. If it is compatible, then we will call `__new__` and `__init__`.

This simplification is important when we consider generic classes. We don't want to use the metaclass `__call__` to instantiate the class's tparams.

Reviewed By: stroxler

Differential Revision: D78759557

fbshipit-source-id: 80c07d54077fccd609955527854f9d9146109462