Disable imports from __pycache__
Summary:
Since we've started allowing `.pyc` imports, we've kind of started allowing people to start importing from `__pycache__` files. For example, `myproject/__pycache__/a.pyc` is importable, even though that's not something that should happen.

We won't suggest it in IDE yet (since we don't work on .pyc files there yet), so excluding it here like this should be enough for now.

Reviewed By: ndmitchell

Differential Revision: D78841658

fbshipit-source-id: 051a57d920d1262c8889a97179a56b6861dcd00e