Support pyx and pyd compiled files for module finding
Summary:
We shouldn't error on importable modules for Cython. This means we should treat `.pyx` files the same as `.pyc` and just treat them as any.

There's also another file type I found with this, `.pyd`, which is basically a DLL-type file. We should exclude it too while we're excluding things.

This should address https://github.com/facebook/pyrefly/issues/703

Reviewed By: ndmitchell

Differential Revision: D78841656

fbshipit-source-id: 04e2b7b045e91ad9edb336f285df9483b01dcc86