Add failing test for Self specialization on overloaded __new__
Summary:
I noticed this bug while reading through the constructor call code. `__new__` is special, because it's a staticmethod so we don't have the `BoundMethod` structure containing the `obj` type. We have a (pretty hacky) setup that extracts the first argument of the call and uses that, but it only applies for the `Target::Function` case, not the `Target::FunctionOverload` case.

I spent a few hours thinking about a fix for this, but ultimately I think we should revisit Self typing more broadly. Going to record this failing test and move on for now.

Reviewed By: rchen152

Differential Revision: D78912235

fbshipit-source-id: f9c6ff878e866108c2872082c551f232200b3874