Filter local variable completion results with terrible matches server side
Summary:
In the previous diff, we will disable auto-import in auto complete if there are local variable results for perf. It might sound reasonable, but in practice it means that auto-import will always be disabled, because we don't filter out local variables with very bad matches, so auto-import will be disabled as long as there is at least one definition in the file.

This diff rectifies the issue by filtering out bad matches on the server side. In this way, we will return fewer results, and at the same time not always disabling auto-import.

Reviewed By: kinto0

Differential Revision: D78923958

fbshipit-source-id: 873b479d3bf33182b3c885d589392b93b3f39a5f