Add testcases for missing binop support
Summary:
For issue https://github.com/facebook/pyrefly/issues/737

The first testcase is failing but we want it to pass.

- mypy fully accepts it
- pyright rejects it. We can add quotes like so https://pyright-play.net/?pyrightVersion=1.1.374&pythonVersion=3.13&strict=true&code=GYJw9gtgBALgngBwJYDsDmUkQWEMoDCAhgDYlEBGJApgDRQDGRAzjPSNQG7WkD68CagCghDcs2ZQAKtRTNcALiFQVUXrxwB3dVAC8jFjAAUxMpRoBtCwCIZc3NfqoYAXXq3Z8kNbdRyECgATIigAD3o4BTCAShFA6mAoYCNQqLsvJyjnaKVVKA5uPgFqFKgAKjLMaKghIA

but it still cannot infer the param or return types.

Seems reasonable that we should try to support this pattern since it blocks PyTorch.

Sam noted that this is related to this issue https://fb.workplace.com/groups/657557118131740/posts/1778016486085792 mentioned by stroxler.

That is related to binding a callable not being turned into a bound method.

I've spent sometime looking at attr.rs but I haven't wrapped my head around exactly when we want to do this. I assume the way we want to bind the function is to use `drop_first_param_of_unbound_callable` and figuring out when to do this.

Reviewed By: stroxler

Differential Revision: D78920731

fbshipit-source-id: 599aabe3d3fb7ce99910bd1ecd7bbe949456d87b