Make matches call infallible
Summary:
This diff makes `Glob::matches()` not return a `Result`, since the fallible operation of constructing the `Pattern` was moved to the constructor instead.

There still is pattern re-construction in `Glob::matches()` that we were ignoring errors from before, and we will continue to ignore it because it's an attempt we make at resolving semantics of the matching library.

Reviewed By: stroxler

Differential Revision: D78908474

fbshipit-source-id: 6348f9ea68db3fe85dff68817c93da62b2693b39