Avoid printing Var unique IDs in chosen overload hover results
Summary:
The chosen overload code path is separate, and was missing the step to prepare for display. Diff also includes a small drive-by refactor for code sharing.

One remaining wrinkle: if the chosen overload is a forall, we don't print the type parameters from the signature. That is, we print `(x: T) -> T` but we should print `[T](x: T) -> T`.

Fixes https://github.com/facebook/pyrefly/issues/739

Reviewed By: SamChou19815

Differential Revision: D78995332

fbshipit-source-id: a447ddf7805d187e755470ab79a0debdcfb684ce