Use `Key::Declaration` instead of `Key::Definition` for `global` and `nonlocal` variables
Summary:
definitionoftest

When we see a statement `global x` and `nonlocal x`, we shouldn't be treating those as definitions of name `x` because that doesn't really model Python semantics well. At runtime these statements merely declare that `x` exists somewhere in the outer scope. The interpreter will NOT really create a `x` for the current scope.

I've fixed similar issue before elsewhere in the codebase D75729810. But I've noticed that we are still creating `Key::Definition()` for globals and nonlocals. This diff proposes that we stop using `Key::Definition()` for them, as they are not definitions. Instead, I introduced a new key `Key::Declaration()` which signals such names.

This change should not have any visible behavior change to the core type checking logic -- type checking currently doesn't really care what kind of keys these are. The point of the change is (1) clarify concepts for future code maintainer, and (2) prepare for gotodef change in D78936649 since the key kind matters a lot for LSP.

Reviewed By: kinto0

Differential Revision: D78936647

fbshipit-source-id: 0c86529ee0ef42307a35937c76b2bd51fac88f7a