Rewrite key-to-definition logic by breaking it into 2 phases
Summary:
noob_rewrite_it_in_rust

I've always had trouble getting a clear understanding of what the function `key_to_intermediate_definition()` does. Given a key `K`, we seem to be looking for its corresponding binding `B`, and invoke `binding_to_intermediate_definition()` on `B`. This can easily give readers an impression that bindings could somehow signal (or correspond to unique) definitions in the code.

But that's the wrong understanding of it! Only certain kinds of keys can signal definitions (I can only find 2 kinds: `Key::Definition`, and `Key::Import`). Bindings merely describe the properties of the language construct that the key refers to, and those properties may reference other keys. The chain formed by key->binding->key->binding->... is often referred to as the "use-def chain", as following this chain would usually lead us to the initial definition that the given programming construct depends on.

Looking from that perspective, it's easy to see that all `binding_to_intermediate_definition()` does is to try to "climb up" the use-def chain. Problem is, it does the climb blindly, reaching all the way till it hits the "end" of the chain. Granted, it is very often the case that the end of the chain must be a binding that corresponds to a def key. But that's not always guaranteed. If we hit a def key in the middle of the chain, or if the end of the chain doesn't correspond to a def key, we'll be computing the wrong result.

This diff rewrites the use-def chain climbing logic in `key_to_intermediate_definition()`/`binding_to_intermediate_definition()`. Instead of climbing all the way till the end, we now look for specific kinds of keys (namely `Key::Definition` and `Key::Import`), and stop the climb after we hit one. This way, we could always guarantee that the climb would always hit a valid def. From there, we compute the final result by looking up the corresponding binding the def key has (resolving any `Forward` along the way), and refine the metadata based on what we see.

Semantically speaking, not much has changed -- we are mostly just doing the same climb, except that now we briefly stop in the middle confirming that we've found a def key. But I believe the new implementation is more modularize (since we now decouple the computation into 2 independent steps, as opposed to having 2 mutually recursive function calling each other), as well as more efficient (each step is smaller and can be easily written iteratively instead of recursively).

Reviewed By: kinto0

Differential Revision: D78936649

fbshipit-source-id: 28a37ad5aed2f2eca78b120a97402b36329f1f49