Fix semantic highlighting for loop iterator variables
Summary:
We had problem semantic-marking loop iterators as variables, e.g.
```
for x in XXX: ...     # `x` is not colored as variable
for x, y in YYY: ...  # both `x` and `y` are not colored as variable
```
In these cases, `Key::Definition(x)` (and `y` respectively) would point to either a `Binding::IteratorValue(...)`, and `Binding::UnpackValue()`. Previously we were not returning the right symbol kinds for these bindings -- which is understandable given that it's not super intuitive to us why these two kinds of bindings would correspond to the variable symbol kind.

But I hope with the explanation in D78936649, we'll start to realize that `Binding::symbol_kind()` doesn't really return the "corresponding symbol kind for the binding". There's no such thing as a "corresponding symbol kind" given a binding. The right understanding of what this function does should be:
- If the corresponding key of a given binding is a def key like `Key::Definition(x)`, what's the most likely symbol kind for name `x`?

Under such interpretation, I think we should definitely return `SymolKind::Variable` for `Binding::IteratorValue(...)`, and `Binding::UnpackValue()`, given how the binding table is currently set up.

One side benefit of  D78936649 is that we automatically get the correct semantic highlighting for comprehension iterators with this diff as well.

Reviewed By: kinto0

Differential Revision: D78936646

fbshipit-source-id: e83ffdbaa17c93ce8f98851ab2ce105bb615b8a4