Add a new `is_declaration` flag to `StaticInfo`
Summary:
i_declare_bankruptcy

In D78936647, we need downstream binding logic to handle real name definitions and mere name declarations (i.e. global and nonlocal) a bit differently. This requires the scope table knowing whether a given name is declared or defined.

This diff introduces a simple flag for such info. We'll always set it to `false` for now to avoid semantic changes. But we'll be utilizing the flag in the next diff.

Reviewed By: kinto0

Differential Revision: D78944028

fbshipit-source-id: 0d4e909df2bcbf0cd6121a8640bf0517425ff4cb