Expand comments around property setter handling
Summary:
I was skimming this code and realized it is pretty confusing, the
types are sort of lying to us here (we treat `property.setter` as having
the same type as `property` except for some metadata).

It all makes sense - this is a much lighter-weight way of implementing
things than using custom types end-to-end - but it's surprising and not
matched to runtime behavior.

There were already some comments in the relevant function decorator and
attribute lookup blocks, but I decided to expand those comments, and also
add one to the relevant `class_field` block since all three of these
components are together defining how Pyrefly handles property setters.

Reviewed By: ndmitchell

Differential Revision: D78908870

fbshipit-source-id: 3b025c4c30ad69e64927c08ec51df4d10621452a