Move phi key creation to MergeItem::new
Summary:
Since I already went down the dubious path of allowing callbacks
because the costs seem to outweigh the benefits, I figured it's
probably worth moving the phi creation into `MergeItem` as well.

My rationale is that this is a tricky operation (we're promising
a binding that we don't create in straight-line control flow) that
is a bit fiddly, so mixing abstraction layers by putting a tricky
bit of merge-on-name logic in the merge-all-names function seems
messy.

Reviewed By: yangdanny97

Differential Revision: D79055910

fbshipit-source-id: 3e6582df9d0650894353b08a7cd3479a62c01bfd