Give the merge item a struct
Summary:
This diff is a no-op logically, it just takes an anonymous tuple
and give the entries names and doc comments by making it a struct.

I expect this to save me 5 minutes of short-term confusion every
single time I read this function, the tuple was pretty crazy
especially because the two `Idx<Key>` values at the beginning of
it are very different semantically (one of them is the idx of an
upstream binding, the other is the idx we need to create a binding
for, which at least in my mind makes them behave very, very
differently).

Reviewed By: yangdanny97

Differential Revision: D79055916

fbshipit-source-id: 6fb7ae80c6b2846ba576f0604d9358273680fb4b