Add `merge_item.add_branch` helper
Summary:
This gets rid of the unnamed `f` that closed over the flow info,
replacing it with a named operation that takes `info` explicitly.

This was another of the bits of the code I had trouble skimming and
regularly had to reread a couple times, I think the control flow
and the "why" of what's going on should be more obvious now.

Reviewed By: yangdanny97

Differential Revision: D79055909

fbshipit-source-id: e66fc068590b129824dd096669fbdb4949ae122b