Find parent project layout type
Summary:
Noticed that `pandas` has an issue where there's a `pyproject.toml` in its installation, which messes up our config finding and search path resolution. We instead find `pandas` to be the search root instead of `site-packages`, so this may be a big cause of missed/improperly inferred types in pandas.

It's kind of difficult to determine when a project is set up like this, and there wasn't anything I really noticed in the `pyproject.toml` that pointed to this setup. What I realized though, is that there's usually some kind of `__init__` file in libraries like this, so we can use that to mark the parent directory as the import root.

This should fix internal and external library imports for pandas, since we can now look in the proper location.

Resolves https://github.com/facebook/pyrefly/issues/749

Reviewed By: ndmitchell

Differential Revision: D79009507

fbshipit-source-id: 67b95980db36a3d6cf30b7a6993df5cd26a4284a