add set_display_mode_to_hover to TypeDisplayContext
Summary: we want the ability to print types differently in hover results. this function sets a flag that will print the type differently (starting in the next diff). this diff is one idea for it, a way to use `TypeDisplayContext` with a method. I'm looking for feedback on this approach. Alternatively, we could adjust the constructor (probably more annoying?)

Reviewed By: ndmitchell

Differential Revision: D79008032

fbshipit-source-id: 8b736eeb960ae1eb66008ad622c68fe8360b313b