Add a comment on short circuit
Summary:
It's not really obvious why this is okay, given that we know
we sometimes need to bind Phi keys.

It winds up being okay because a loop always has at least two
branches, and we don't use speculative phi on other branching
control structures, but that really deserves an explicit
mention because this sort of thing can be panic-prone.

Reviewed By: yangdanny97

Differential Revision: D79055911

fbshipit-source-id: 224bf2b5b1675e0905a58e6f23a0d62949f80af7