Document the bind-current helpers a little better
Summary:
The name `bind_definition_current` confused me, I think at some point it
was renamed from `bind_definition_idx`, but it's not actually binding a
definition (the `current` is an input), so the "definition" part of the name
was misleading.

This diff renames it to `bind_current_as`

While I'm addressing that, it's useful to add comments to both `bind_current_as`
and `bind_current` explaining how they differ and what the caller is expected to
do if they use `bind_current`.

Reviewed By: ndmitchell

Differential Revision: D79055912

fbshipit-source-id: 8446b319442b5281096aac83ca3506e6a6969d26