Rename and simplify some branch handling
Summary:
We have terminology for a terminated branch, it's called a terminated
branch. Why invent a second set of terms "visible" and "hidden" that
mean exactly the same thing, but less clearly.

Now we don't need a comment that defines new terms to be exactly what
the existing terms were.

I also think it's clearer to just define `branches` to be what we are
going to merge instead of mutating `visible_branches`. This doesn't
even cost us a clone, and it makes the names and control flow clearer
in my opinion.

Reviewed By: yangdanny97

Differential Revision: D79055907

fbshipit-source-id: 2d2d92a64eff179d4dcbb20df95e49449d7a4d5f